pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pytia"
version = "0.1.0"
description = "PyTIA: voxel-wise time-integrated activity (TIA) maps from multi-timepoint PET/SPECT"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [{name = "PyTIA contributors"}]
dependencies = [
  "numpy>=1.26",
  "scipy>=1.12",
  "nibabel>=5.2",
  "PyYAML>=6.0.1",
  "scikit-image>=0.22",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.6",
]
# Optional plotting for scripts (not required)
viz = [
  "matplotlib>=3.8",
]

[project.scripts]
pytia = "pytia.cli:main"

[tool.setuptools.packages.find]
include = ["pytia*"]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]